refactor(fsd): utils の純粋関数を shared/lib へ移設 (#356)
A2 Lane2 PR#1(shared foundation)の一部。src/utils/ は FSD レイヤ外なので
shared/lib へ。format(yen/formatDate/formatDateTime/daysOverdue)と
fiscal(会計年度計算)は context-free な純粋関数＝shared/lib の定義に合致。

- src/utils/format.ts → src/shared/lib/format.ts
- src/utils/fiscal.ts(+test) → src/shared/lib/
- @/utils/{format,fiscal} の import 9箇所を @/shared/lib/ に書換。
- format.ts 冒頭に逸脱注記(#358): 規約は formatter を shared/lib 禁止で
  nene2-i18n/format が正準だが、その sub-path が W0b 未実装(0.1.0 exports
  は "." のみ)＝移設先が存在しないための経過措置(hub 裁定 07-17)。

挙動不変(移設と import 書換のみ)。type-check 緑・62 tests 緑・@/utils 残存 0。
